Fix warning (#13571)
* remove optional and resource plugin

* fix some rbac type warning

* suppress unchecked and rawtypes

* suppress compute warning

* fix app service warning

* fix cosmos warning

* fix aks warning

* fix monitor warning

* change msi handler to generic type rather than suppress rawtypes

* change ExpressRouteCircuitPeering to generic type

+++ sdk/resourcemanager/azure-resourcemanager-compute/src/test/java/com/azure/resourcemanager/compute/VirtualMachineScaleSetOperationsTests.java
@@ -207,7 +207,7 @@ public class VirtualMachineScaleSetOperationsTests extends ComputeManagementTest
         Assertions.assertNotNull(extension.publicSettingsAsJsonString());
         // Update VMSS capacity
         //
-        int newCapacity = (int) (scaleSet.capacity() + 1);
+        int newCapacity = scaleSet.capacity() + 1;
         virtualMachineScaleSet.update().withCapacity(newCapacity).apply();
         // Validate extensions after update
         //
